user_profile/serializers.py
from rest_framework import serializers
from django.contrib.auth.models import User
from user_profile.models import UserProfile


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['username', 'password']
        extra_kwargs = {'password': {'write_only': True}}

    def create(self, validated_data):
        password = validated_data.pop('password')
        user = User(**validated_data)
        user.set_password(password)
        user.save()
        UserProfile.objects.create(user=user)
        return user


class UserProfileSerializer(serializers.ModelSerializer):
    # icon_url = serializers.SerializerMethodField()

    class Meta:
        model = UserProfile
        fields = ['name', 'surname', 'icon', 'description']
        extra_kwargs = {'user__username': {'read_only': True}}
